Added a dedicated fence class, base command queue is now a level 2 interface, added support for right handed ortho / perspective

+++ Projects/RendererDX12/Source/Level 2/DX12Mesh.cpp
@@ -36,8 +36,8 @@ void DX12Mesh::CreateVertexBuffer(DX12CommandList* commandList, void* vertexData
 	commandList->GetBase()->GetInterface()->ResourceBarrier(1, &barrier);
 
 	m_vertexView.BufferLocation = m_vertexBuffer.Get()->GetGPUVirtualAddress();
-	m_vertexView.SizeInBytes = totalVertexSize;
-	m_vertexView.StrideInBytes = sizeOfVertex;
+	m_vertexView.SizeInBytes = static_cast<UINT>(totalVertexSize);
+	m_vertexView.StrideInBytes = static_cast<UINT>(sizeOfVertex);
 
 	m_vertexCount = vertexCount;
 }
